* 'master' of ssh://git.isartintra.com:2424/2019/GDP_2022_GA_2022_MS_2021_Platfor/SEQUOIA: (26 commits)
  Block inputs oflogin screen
  Improve ScoreObject and fix bug on displaying player score
  Good retry for build
  Done
  Leaderboard is done ! (in future will make the flow better)
  Well displayed in every counditions, change between levels is missing
  Scores well get on Leaderboard
  Server minor modfification
  Advance getting scores
  Advance client getting scores
  Register scores done
  Fix error (user_id was used instead of results[0].user_id)
  Test server id + token response to sign roads
  Prepare score rregistering
  Clean UI Screens code
  Hide password field and don't ask twice for logging when it first succeeds
  Prepare leaderboard display + webClient not implemented exception on event LevelManager.OnWin isn't thrown
  Login complete
  Start really login in
  Use transparent image to prevent clicking behind loginScreen
  ...

+++ Assets/_Platformer/Scripts/LevelObjects/Collectibles/ScoreCollectible.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 
 namespace Com.IsartDigital.Platformer.LevelObjects.Collectibles
 {
-    public delegate void ScoreCollectibleEventHandler(float score);
+    public delegate void ScoreCollectibleEventHandler(int score);
 
     public class ScoreCollectible : ACollectible {
         private static List<ScoreCollectible> _list = new List<ScoreCollectible>();
@@ -16,7 +16,7 @@ namespace Com.IsartDigital.Platformer.LevelObjects.Collectibles
 
         public event ScoreCollectibleEventHandler OnCollected;
 
-        [SerializeField] private float score = 1;
+        [SerializeField] private int score = 1;
 
         private void Awake()
         {
